New AutoDraft feature + template adjustment
Added a new feature called "AutoDraft". It saves your messages in the browser's Local Storage as "draft", so in case your browser or your PC crashes, the message that you were writing for an hour will be restored if you visit the thread. This is only for replies (Quick Reply and the general Reply), not for editing posts, or creating new threads. (yet?)
Once you send your reply, the draft will disappear from your Local Storage, meaning it won't be restored again.
You can also have a "timer", so if you have an unused draft in your Local Storage that's there for a period of time (that you can specify via HFX config), once the timer elapses, it will be purged automatically.

Also adjusted the "Feature" template so it won't conflict with ESLint anymore. (unnecessary semicolon by default)